slice2py simplification, comments, and cleanup
