+++ cpp/src/slice2py/PythonUtil.cpp
@@ -2761,21 +2761,27 @@ namespace
 }
 
 vector<Slice::Python::PythonCodeFragment>
-Slice::Python::dynamicCompile(const vector<string>& files, const vector<string>& cppArgs, bool debug)
+Slice::Python::dynamicCompile(const vector<string>& files, const vector<string>& preprocessorArgs, bool debug)
 {
+    // The package visitor is reused to compile all Slice files so that it collects the definitions for all Slice files
+    // contributing to a given package.
     PackageVisitor packageVisitor;
+
+    // The import visitor is reused to collect the imports from all generated Python modules, which are later used to
+    // compute the order in which Python modules should be evaluated.
     ImportVisitor importVisitor;
+
     // The list of code fragments generated by the code visitor.
     vector<PythonCodeFragment> fragments;
 
     for (const auto& fileName : files)
     {
-        PreprocessorPtr preprocessor = Preprocessor::create("IcePy", fileName, cppArgs);
+        PreprocessorPtr preprocessor = Preprocessor::create("IcePy", fileName, preprocessorArgs);
         FILE* cppHandle = preprocessor->preprocess(true, "-D__SLICE2PY__");
 
         if (cppHandle == nullptr)
         {
-            throw runtime_error("Failed to preprocess Slice files");
+            throw runtime_error("Failed to preprocess Slice file: " + fileName);
         }
 
         UnitPtr unit = Unit::createUnit("python", debug);
@@ -2784,7 +2790,7 @@ Slice::Python::dynamicCompile(const vector<string>& files, const vector<string>&
         if (parseStatus == EXIT_FAILURE)
         {
             unit->destroy();
-            throw runtime_error("Failed to parse Slice files");
+            throw runtime_error("Failed to parse Slice file: " + fileName);
         }
 
         parseAllDocComments(unit, Slice::Python::pyLinkFormatter);
@@ -2799,10 +2805,8 @@ Slice::Python::dynamicCompile(const vector<string>& files, const vector<string>&
             importVisitor.getAllImportNames()};
         unit->visit(&codeVisitor);
 
-        for (const auto& fragment : codeVisitor.codeFragments())
-        {
-            fragments.push_back(fragment);
-        }
+        const vector<PythonCodeFragment>& newFragments = codeVisitor.codeFragments();
+        fragments.insert(fragments.end(), newFragments.begin(), newFragments.end());
         unit->destroy();
     }
 
@@ -2881,15 +2885,14 @@ Slice::Python::dynamicCompile(const vector<string>& files, const vector<string>&
         if (fragments.size() == fragmentsSize)
         {
             assert(false);
-            throw runtime_error(
-                "Circular dependency detected in Slice files. Please check the imports and package definitions.");
+            throw runtime_error("Circular dependency detected in Slice files.");
         }
     }
 
     // Add fragments for the package-index files.
     for (const auto& [name, imports] : packageVisitor.imports())
     {
-        string packageName = name; // Remove the trailing dot.
+        string packageName = name;
         // The pop_back call removes the last dot from the package name.
         packageName.pop_back();
         BufferedOutput out;
