+++ python/modules/IcePy/Makefile.mk
@@ -11,7 +11,6 @@ IcePy_dependencies      := IceDiscovery IceLocatorDiscovery Ice
 IcePy_libs              := mcpp
 IcePy_extra_sources     := $(wildcard $(top_srcdir)/cpp/src/Slice/*.cpp) \
                            $(top_srcdir)/cpp/src/slice2py/PythonUtil.cpp \
-                           $(top_srcdir)/cpp/src/slice2py/Python.cpp     \
 
 #
 # On the default platform, always write the module in the python directory.
